Remove delay part 2
Using feedkeys to run a Lua function doesn't feel quite right, but
this must be run only after the cursor is correctly placed, which is
done via feedkeys too.  If the CursorMoved event is triggered before the
cursor is in place, the snippet state might be cleared, because we track
the cursor position to do it.