Add custom expand options
Based on the ideas behind the expand options: i,A,b,w to limit
triggering an expansion to certain scenarios, this commit allows for
user-defined expand options. The user definition in the `setup` function
should define the option letter as a key and a function to be called with
no arguments to determine if expansion is possible.

The custom expand options are currently given priority over i,A,b,w to
ensure they are not ignored.

Test for custom expand option (commented line)