Use non-deprecated version of `vim.validate()`
Since NeoVim 0.11, `vim.validate()` has a new signature which uses less
tables, is more performant and easier to read. The old signature is
considered deprecated and is targeted for removal in NeoVim 1.0. Even
though we won't see NeoVim 1.0 soon, it's somewhat annoying to trigger
deprecation warnings each time the plugin is loaded.

This patch adds a condition to two places where `vim.validate()` is
used, using the new signature on NeoVim 0.11 and above.